Spark,Core: Refactor Delete OrphanFiles by moving common code from Spark to core
